Ownership and authentication

+++ src/jmp.py
@@ -30,6 +30,7 @@ class Link(BASE):
     link_id = Column(Integer, primary_key=True, autoincrement=True)
     shorty = Column(String(256), nullable=False, unique=True)
     longfellow = Column(String(2048), nullable=False)
+    owner = Column(String(len(sha256("").hexdigest())), nullable=False)
 
 APP = Flask(__name__)
 
@@ -41,6 +42,8 @@ RESERVED_SHORTS = ("api",)
 SUPPORTED_SHORT_RE = r"^\w+$"
 MAX_LONGFELLOW_SIZE = 2048
 MAX_SHORT_SIZE = 140
+
+ENTRY_UUID_HASH = "LOLMIGHTYSALTYINHERE!" #TODO: move this to a protected conf
 
 DBSESSION = sessionmaker(bind=ENGINE)
 
@@ -59,9 +62,18 @@ def req_auth_api(func):
         username = request.environ.get("X-WEBAUTH-USERNAME", None)
 
         if entry_uuid == None or username == None:
-            return "NOPE" #TODO: do something more intelligent...
+            return json.dumps([{"success" : False,
+                "error" : "Authentication required"}])
         return func(*args, **kwargs)
     return auth_decor
+
+def hash_and_salt(plain, salt=ENTRY_UUID_HASH):
+    """
+    Designed to be used for entry-uuid hashing, return a sha256 hash of 'plain'
+    after applying 'salt'
+    """
+
+    return sha256("{0}|{1}".format(plain, salt)).hexdigest()
 
 def _verify_long(longfellow):
     """
@@ -121,15 +133,21 @@ def redir(short):
 #    return str(request.cookies.items())
 
 @APP.route("/api/new")
+@req_auth_api
 def add_link():
     """
     A method to add an entry to the links table.
     """
 
-    #XXX: AUTHENTICATE FOR THIS METHOD
     longfellow = request.args.get("long", None)
     shorty = request.args.get("short", None)
-    #owner = #TODO: ... however I get the user's identity ...
+    entry_uuid = request.environ.get("X-WEBAUTH-ENTRYUUID", None)
+
+    if entry_uuid == None:
+        return json.dumps([{"success" : False,
+            "error" : "That's weird. You lack an entry-uuid..."}])
+
+    owner = hash_and_salt(entry_uuid)
 
     if not _verify_short(shorty):
         return json.dumps([{"success" : False,
@@ -145,7 +163,7 @@ def add_link():
 
     try:
         session = DBSESSION()
-        new_link = Link(shorty=shorty, longfellow=longfellow)
+        new_link = Link(shorty=shorty, longfellow=longfellow, owner=owner)
         session.add(new_link)
         session.commit()
 
@@ -155,30 +173,51 @@ def add_link():
             "error" : exception.args}])
 
 @APP.route("/api/delete")
+@req_auth_api
 def rm_link():
     """
     A method to remove an entry from the links table.
     """
 
-    #XXX: AUTHENTICATE FOR THIS METHOD
     longfellow = request.args.get("long", None)
     shorty = request.args.get("short", None)
-    #user = #TODO: ... however I get the user's identity ...
+
     if shorty == None or longfellow == None:
         return json.dumps([{"success" : False,
             "error" : "Incomplete request"}])
 
+    entry_uuid = request.environ.get("X-WEBAUTH-ENTRYUUID", None)
+
+    if entry_uuid == None:
+        return json.dumps([{"success" : False,
+            "error" : "That's weird. You lack an entry-uuid..."}])
+
+    owner = hash_and_salt(entry_uuid)
+
     try:
         session = DBSESSION()
-        #TODO: verify that the user owns this link
-        #    : ... this may be possible by simply including the username
-        #    :     in the schema and querying based on that
 
         matching_links = session.query(Link).filter_by(
                 shorty=shorty, longfellow=longfellow)
 
+        link_deleted = False
+        at_least_one_link = False
+
         for link in matching_links:
-            session.delete(link)
+            at_least_one_link = True
+            if link.owner == owner: #TODO: also check if user is admin
+                session.delete(link)
+                link_deleted = True
+
+        if not at_least_one_link:
+            #Just in case we ever want to extend the schema to allow separation
+            #of privately and publicly shared links
+            return json.dumps([{"success" : False,
+                "error" : "Link not found"}])
+
+        if not link_deleted:
+            return json.dumps([{"success" : False,
+                "error" : "Keep your hands to yourself"}])
 
         session.commit()
         return json.dumps([{"success" : True}])
